test: split solder paste tests one per file, use standard pcb snapshot
